+++ app/Models/Admin/Ratings.php
@@ -252,13 +252,16 @@ class Ratings extends AppModel
         $images = $staff->getResizeImagesAttribute();
     	if($staff->delete())
         {
-            foreach($images as $img)
+            if($images)
             {
-                foreach($img as $i)
+                foreach($images as $img)
                 {
-                    if($i && is_dir(public_path($i)) && file_exists(public_path($i)))
+                    foreach($img as $i)
                     {
-                        unlink(public_path($i));
+                        if($i && is_dir(public_path($i)) && file_exists(public_path($i)))
+                        {
+                            unlink(public_path($i));
+                        }
                     }
                 }
             }
