+++ app/Models/Admin/StaffDocuments.php
@@ -256,13 +256,16 @@ class StaffDocuments extends AppModel
         $images = $product->getResizeImagesAttribute();
     	if($product->delete())
         {
-            foreach($images as $img)
+            if($images)
             {
-                foreach($img as $i)
+                foreach($images as $img)
                 {
-                    if($i && is_dir(public_path($i)) && file_exists(public_path($i)))
+                    foreach($img as $i)
                     {
-                        unlink(public_path($i));
+                        if($i && is_dir(public_path($i)) && file_exists(public_path($i)))
+                        {
+                            unlink(public_path($i));
+                        }
                     }
                 }
             }
